Update test_runlib.py
Fixes issue in-toto#365:

Previously, 3 tests from test_runlib.py failed on a Windows platform, because symlinks could not be dealt with if developer mode wasn't activated. The code now has added functionality to be able to check the platform of the user's system. When necessary (like in the methods that use symlinks), if the platform is windows, test_runlib.py raises an error that asks the user if the developer mode is activated.